ART: Fix warnings
Fix Wconstant-conversion and Wundefined-var-template warnings.

Reverts commit df53be273509dd43725870fb20a2c7d71f7fbfd3.

Bug: 28149048
Bug: 29823425
Test: m
Test: m test-art-host

+++ compiler/optimizing/code_generator_arm.cc
@@ -2522,7 +2522,7 @@ void CodeGeneratorARM::GenerateFrameEntry() {
   if (GetGraph()->HasShouldDeoptimizeFlag()) {
     // Initialize should_deoptimize flag to 0.
     __ mov(IP, ShifterOperand(0));
-    __ StoreToOffset(kStoreWord, IP, SP, -kShouldDeoptimizeFlagSize);
+    __ StoreToOffset(kStoreWord, IP, SP, -static_cast<int32_t>(kShouldDeoptimizeFlagSize));
   }
 
   int adjust = GetFrameSize() - FrameEntrySpillSize();
